Data collection module should be finished now. + Proper type checking of Object arrays given to the MetricDefinitions. + Links should now be operational. Links between Locations are standard, links with an Ibis at either end are optional (on demand).

+++ src/junglevision/gathering/impl/Location.java
@@ -466,11 +466,11 @@ public class Location extends Element implements junglevision.gathering.Location
 							metric.setValue(MetricModifier.MIN, outputtype, min);
 						}						
 					} else { //We are MetricOutput.N
-						int total  = 0, max = 0, min = 1000000;
+						long total  = 0, max = 0, min = 1000000;
 						
 						//First, we gather our own metrics
 						for (junglevision.gathering.Ibis ibis : ibises) {							
-							int ibisValue = (Integer) ibis.getMetric(desc).getValue(MetricModifier.NORM, outputtype);
+							long ibisValue = (Long) ibis.getMetric(desc).getValue(MetricModifier.NORM, outputtype);
 							
 							total += ibisValue ;
 							
@@ -480,7 +480,7 @@ public class Location extends Element implements junglevision.gathering.Location
 						
 						//Then we add the metric values of our child locations					
 						for (junglevision.gathering.Location child : children) {
-							int childValue = (Integer) child.getMetric(desc).getValue(MetricModifier.NORM, outputtype);
+							long childValue = (Long) child.getMetric(desc).getValue(MetricModifier.NORM, outputtype);
 							
 							total += childValue;
